services/distributed/requirements.txt
# 分布式滤波服务依赖
numpy>=1.21.0
scipy>=1.7.0
kafka-python>=2.0.2
websockets>=10.0
